Move listener and split some declarations to avoid duplicates and to use delegation
