+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/listener/CompositeSocketChannelListener.kt
@@ -1,4 +1,4 @@
-package co.omisego.omisego.websocket
+package co.omisego.omisego.websocket.listener
 
 /**
  * OmiseGO
@@ -8,6 +8,7 @@ package co.omisego.omisego.websocket
  */
 
 import co.omisego.omisego.model.APIError
+import co.omisego.omisego.websocket.SocketChannelListener
 
 class CompositeSocketChannelListener(
     private val listeners: MutableSet<SocketChannelListener> = linkedSetOf()
