+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/listener/SocketChannelListenerSet.kt
@@ -1,0 +1,8 @@
+package co.omisego.omisego.websocket.listener
+
+import co.omisego.omisego.websocket.SocketChannelListener
+
+interface SocketChannelListenerSet {
+    fun addChannelListener(channelListener: SocketChannelListener)
+    fun removeChannelListener(channelListener: SocketChannelListener)
+}
