+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/listener/SocketConnectionListenerSet.kt
@@ -1,0 +1,8 @@
+package co.omisego.omisego.websocket.listener
+
+import co.omisego.omisego.websocket.SocketConnectionListener
+
+interface SocketConnectionListenerSet {
+    fun addConnectionListener(connectionListener: SocketConnectionListener)
+    fun removeConnectionListener(connectionListener: SocketConnectionListener)
+}
