+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/listener/SocketCustomEventListenerSet.kt
@@ -1,0 +1,8 @@
+package co.omisego.omisego.websocket.listener
+
+import co.omisego.omisego.websocket.SocketCustomEventListener
+
+interface SocketCustomEventListenerSet {
+    fun addCustomEventListener(customEventListener: SocketCustomEventListener)
+    fun removeCustomEventListener(customEventListener: SocketCustomEventListener)
+}
